federal_booking_bot_system/requirements.txt
aiogram

requests

selenium
python-dotenv

chromedriver-autoinstaller